python_director/requirements.txt
google-genai
firebase-admin
python-dotenv
pytest
pydantic
openai
fastapi
uvicorn
python-multipart
